第22天，完结

+++ pinyougou-model/src/main/java/com/pinyougou/model/Item.java
@@ -39,7 +39,7 @@ public class Item implements Serializable {
      */
     @Column(name = "price")
     @Field(value = "item_price")
-    private Double price;
+    private BigDecimal price;
 
     @Column(name = "stock_count")
     private Integer stockCount;
@@ -198,7 +198,7 @@ public class Item implements Serializable {
      *
      * @return price - 商品价格，单位为：元
      */
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
@@ -207,7 +207,7 @@ public class Item implements Serializable {
      *
      * @param price 商品价格，单位为：元
      */
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
